chore: removed default metadata logic from LocaleLayout and move it into home page file
