+++ lang/es.json
@@ -1,9 +1,4 @@
 {
-  "metadata": {
-    "title": "Bachacode Developers | Desarrollo & Consultoría Web",
-    "description": "Bachacode Developers se especializa en desarrollo web a medida, ofreciendo sitios estáticos, aplicaciones web y WordPress, ajustados a tus necesidades.",
-    "keywords": "desarrollo, programación, wordpress, php, laravel, soporte"
-  },
   "navigation": {
     "home": "Inicio",
     "about_us": "Sobre Nosotros",
@@ -23,6 +18,10 @@
     "button": "Regresar al Inicio"
   },
   "home": {
+    "metadata": {
+      "title": "Desarrollo & Consultoría Web",
+      "description": "Bachacode Developers se especializa en desarrollo web a medida, ofreciendo sitios estáticos, aplicaciones web y WordPress, ajustados a tus necesidades."
+    },
     "hero": {
       "title": "Construimos <websites>sitios web</websites> increibles para negocios pequeños",
       "subtitle": "con nosotros tu nuevo proyecto se hará realidad",
