+++ src/app/[locale]/layout.tsx
@@ -1,9 +1,8 @@
 import { routing } from "@/i18n/routing";
-import { getTranslations, setRequestLocale } from "next-intl/server";
+import { setRequestLocale } from "next-intl/server";
 import { notFound } from "next/navigation";
 import MainHeader from "@/components/layout/MainHeader";
 import MainFooter from "@/components/layout/MainFooter";
-import { Metadata } from "next";
 import { hasLocale, NextIntlClientProvider } from "next-intl";
 import ProgressBarProvider from "@/components/layout/ProgressBarProvider";
 import { Analytics } from "@vercel/analytics/next";
@@ -14,25 +13,6 @@ const roboto = Roboto({ weight: "500", subsets: ["latin"], display: "swap" });
 
 export function generateStaticParams() {
   return routing.locales.map((locale) => ({ locale }));
-}
-
-export async function generateMetadata({
-  params,
-}: {
-  params: { locale: string };
-}): Promise<Metadata> {
-  const { locale } = await params;
-
-  const t = await getTranslations({
-    locale: hasLocale(routing.locales, locale) ? locale : routing.defaultLocale,
-    namespace: "metadata",
-  });
-
-  return {
-    title: t("title"),
-    description: t("description"),
-    keywords: t("keywords"),
-  };
 }
 
 export default async function LocaleLayout({
