+++ src/app/[locale]/page.tsx
@@ -21,10 +21,39 @@ import bachacodeBig from "../../../public/images/bachacode-big.png";
 import HomePageCounter from "@/components/sections/HomePageCounter";
 import PageSectionWrapper from "@/components/layout/PageSectionWrapper";
 import ContactSection from "@/components/sections/ContactSection";
-import { useTranslations } from "next-intl";
+import { hasLocale, useTranslations } from "next-intl";
 import { Button } from "@/components/ui/button";
 import { Link } from "@/i18n/navigation";
 import UnderlinedText from "@/components/common/UnderlinedText";
+import { Metadata } from "next";
+import { getTranslations } from "next-intl/server";
+import { routing } from "@/i18n/routing";
+
+export async function generateMetadata({
+  params,
+}: {
+  params: { locale: string };
+}): Promise<Metadata> {
+  const { locale } = await params;
+
+  const t = await getTranslations({
+    locale: hasLocale(routing.locales, locale) ? locale : routing.defaultLocale,
+    namespace: "home.metadata",
+  });
+
+  return {
+    title: t("title"),
+    description: t("description"),
+    alternates: {
+      canonical: "/",
+      languages: {
+        en: "/",
+        es: "/es",
+        "x-default": "/",
+      },
+    },
+  };
+}
 
 export default function Home() {
   const t = useTranslations("home");
